Separate SemantikonDiGraph
